+++ src/otx/backend/native/engine/__init__.py
@@ -1,0 +1,8 @@
+"""Native engine."""
+
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from .engine import OTXEngine
+
+__all__ = ["OTXEngine"]
