+++ src/otx/backend/native/engine/adaptive_bs/adaptive_bs_api.py
@@ -21,13 +21,13 @@ from .bs_search_algo import BsSearchAlgo
 if TYPE_CHECKING:
     from lightning import LightningModule, Trainer
 
-    from otx.engine.engine import Engine
+    from otx.backend.native.engine import OTXEngine
 
 logger = logging.getLogger(__name__)
 
 
 def adapt_batch_size(
-    engine: Engine,
+    engine: OTXEngine,
     not_increase: bool = True,
     callbacks: list[Callback] | Callback | None = None,
     **train_args,
@@ -38,7 +38,7 @@ def adapt_batch_size(
     If not_increase is False, increase batch size to use most of GPU memory.
 
     Args:
-        engine (Engine): engine instnace.
+        engine (OTXEngine): engine instnace.
         not_increase (bool) : Whether adapting batch size to larger value than default value or not.
         callbacks (list[Callback] | Callback | None, optional): callbacks used during training. Defaults to None.
     """
@@ -89,7 +89,7 @@ def _adjust_train_args(train_args: dict[str, Any]) -> dict[str, Any]:
     return train_args
 
 
-def _train_model(bs: int, engine: Engine, callbacks: list[Callback] | Callback | None = None, **train_args) -> None:
+def _train_model(bs: int, engine: OTXEngine, callbacks: list[Callback] | Callback | None = None, **train_args) -> None:
     if bs <= 0:
         msg = f"Batch size should be greater than 0, but {bs} is given."
         raise ValueError(msg)
@@ -162,7 +162,7 @@ def _scale_batch_reset_params(trainer: Trainer, steps_per_trial: int) -> None:
         trainer.limit_val_batches = steps_per_trial
 
 
-def _apply_new_batch_size(engine: Engine, new_batch_size: int) -> None:
+def _apply_new_batch_size(engine: OTXEngine, new_batch_size: int) -> None:
     origin_bs = engine.datamodule.train_subset.batch_size
     if new_batch_size == origin_bs:
         return
