+++ src/otx/cli/utils/help_formatter.py
@@ -162,10 +162,10 @@ def render_guide(subcommand: str | None = None) -> list:
     """
     if subcommand is None:
         return []
-    from otx.engine import Engine
+    from otx.backend.native.engine import OTXEngine
 
     contents: list[Panel | Markdown] = [Markdown(INTRO_MARKDOWN)]
-    target_command = getattr(Engine, subcommand)
+    target_command = getattr(OTXEngine, subcommand)
     cli_usage = get_cli_usage_docstring(target_command)
     if cli_usage is not None:
         cli_usage += f"\n{VERBOSE_USAGE.format(subcommand=subcommand)}"
