+++ src/otx/types.py
@@ -1,0 +1,15 @@
+"""Typing hints for OTX."""
+
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
+
+from typing import Any
+
+from otx.core.model.base import OTXModel
+
+METRICS = Any  # TODO(ashwinvaidya17): Temporary till metrics is properly defined
+ANNOTATIONS = Any  # TODO(ashwinvaidya17): Temporary till annotations is properly defined
+MODEL = OTXModel  # TODO(ashwinvaidya17): Temporary till model is properly defined
+DATA = Any  # TODO(ashwinvaidya17): Temporary till data is properly defined
