+++ tests/integration/cli/test_auto_configuration.py
@@ -6,8 +6,8 @@ from pathlib import Path
 
 import pytest
 
+from otx.backend.native.engine.utils.auto_configurator import DEFAULT_CONFIG_PER_TASK
 from otx.core.types.task import OTXTaskType
-from otx.engine.utils.auto_configurator import DEFAULT_CONFIG_PER_TASK
 from tests.utils import run_main
