+++ tests/integration/cli/test_cli.py
@@ -9,8 +9,8 @@ import pytest
 import torch
 import yaml
 
+from otx.backend.native.engine.utils.auto_configurator import DEFAULT_CONFIG_PER_TASK
 from otx.core.types.task import OTXTaskType
-from otx.engine.utils.auto_configurator import DEFAULT_CONFIG_PER_TASK
 from tests.utils import ExportCase2Test, run_main
