+++ tests/perf_v2/benchmark.py
@@ -15,8 +15,8 @@ from typing import Any, Literal
 import pandas as pd
 from jsonargparse import ArgumentParser, Namespace
 
+from otx.backend.native.engine import OTXEngine
 from otx.core.types.task import OTXTaskType
-from otx.engine import Engine
 from tests.perf_v2 import CRITERIA_COLLECTIONS, DATASET_COLLECTIONS, MODEL_COLLECTIONS, summary
 from tests.perf_v2.utils import (
     Criterion,
@@ -307,7 +307,7 @@ class Benchmark:
         dataset_info: DatasetInfo,
         work_dir: Path,
         subcommand: SubCommand,
-    ) -> tuple[Engine, dict[str, Any]]:
+    ) -> tuple[OTXEngine, dict[str, Any]]:
         """Initialise engine with given model and dataset settings.
 
         Args:
@@ -319,7 +319,7 @@ class Benchmark:
             Engine: Initialised engine
         """
 
-        engine = Engine(
+        engine = OTXEngine(
             model=model_info.name,
             data_root=self.data_root / dataset_info.path,
             work_dir=work_dir,
@@ -331,7 +331,7 @@ class Benchmark:
         # Instantiate Train Arguments
         engine_parser = ArgumentParser()
         arguments = engine_parser.add_method_arguments(
-            Engine,
+            OTXEngine,
             subcommand.value,
             skip={"accelerator", "devices"},
             fail_untyped=False,
