+++ tests/unit/engine/adaptive_bs/test_bs_search_algo.py
@@ -5,8 +5,12 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from otx.engine.adaptive_bs import bs_search_algo as target_file
-from otx.engine.adaptive_bs.bs_search_algo import BsSearchAlgo, _get_max_memory_reserved, _get_total_memory_size
+from otx.backend.native.engine.adaptive_bs import bs_search_algo as target_file
+from otx.backend.native.engine.adaptive_bs.bs_search_algo import (
+    BsSearchAlgo,
+    _get_max_memory_reserved,
+    _get_total_memory_size,
+)
 
 
 @pytest.fixture()
