+++ tests/unit/engine/test_engine.py
@@ -11,21 +11,21 @@ from pytest_mock import MockerFixture
 
 from otx.algo.classification.multiclass_models import EfficientNetMulticlassCls
 from otx.algo.classification.multiclass_models.torchvision_model import TVModelMulticlassCls
+from otx.backend.native.engine import OTXEngine
 from otx.core.model.base import DataInputParams, OTXModel, OVModel
 from otx.core.types.export import OTXExportFormatType
 from otx.core.types.label import NullLabelInfo
 from otx.core.types.precision import OTXPrecisionType
-from otx.engine import Engine
 from tests.unit.core.utils.test_utils import get_dummy_ov_cls_model
 
 
 @pytest.fixture()
-def fxt_engine(tmp_path) -> Engine:
+def fxt_engine(tmp_path) -> OTXEngine:
     recipe_path = "src/otx/recipe/classification/multi_class_cls/tv_mobilenet_v3_small.yaml"
     data_root = "tests/assets/classification_dataset"
     task_type = "MULTI_CLASS_CLS"
 
-    return Engine.from_config(
+    return OTXEngine.from_config(
         config_path=recipe_path,
         data_root=data_root,
         task=task_type,
@@ -36,11 +36,11 @@ def fxt_engine(tmp_path) -> Engine:
 class TestEngine:
     def test_constructor(self, tmp_path) -> None:
         with pytest.raises(RuntimeError):
-            Engine(work_dir=tmp_path)
+            OTXEngine(work_dir=tmp_path)
 
         # Check auto-configuration
         data_root = "tests/assets/classification_dataset"
-        engine = Engine(work_dir=tmp_path, data_root=data_root)
+        engine = OTXEngine(work_dir=tmp_path, data_root=data_root)
         assert engine.task == "MULTI_CLASS_CLS"
         assert engine.datamodule.task == "MULTI_CLASS_CLS"
         assert isinstance(engine.model, EfficientNetMulticlassCls)
@@ -54,7 +54,7 @@ class TestEngine:
 
         # Create engine with no data_root
         with pytest.raises(ValueError, match="Given model class (.*) requires a valid label_info to instantiate."):
-            _ = Engine(work_dir=tmp_path, task="MULTI_CLASS_CLS")
+            _ = OTXEngine(work_dir=tmp_path, task="MULTI_CLASS_CLS")
 
     @pytest.fixture()
     def mock_datamodule(self, mocker):
@@ -71,7 +71,7 @@ class TestEngine:
 
     def test_model_init(self, tmp_path, mock_datamodule):
         data_root = "tests/assets/classification_dataset"
-        engine = Engine(work_dir=tmp_path, data_root=data_root)
+        engine = OTXEngine(work_dir=tmp_path, data_root=data_root)
 
         assert engine._model.data_input_params == DataInputParams((1234, 1234), (0.0, 0.0, 0.0), (1.0, 1.0, 1.0))
         assert engine._model.label_info.num_classes == 4321
@@ -341,7 +341,7 @@ class TestEngine:
             "data.test_subset.subset_name": "TESTING",
         }
 
-        engine = Engine.from_model_name(
+        engine = OTXEngine.from_model_name(
             model_name=model_name,
             data_root=data_root,
             task=task_type,
@@ -354,7 +354,7 @@ class TestEngine:
         assert engine.datamodule.test_subset.subset_name == "TESTING"
 
         with pytest.raises(FileNotFoundError):
-            engine = Engine.from_model_name(
+            engine = OTXEngine.from_model_name(
                 model_name="wrong_model",
                 data_root=data_root,
                 task=task_type,
@@ -372,7 +372,7 @@ class TestEngine:
             "data.test_subset.subset_name": "TESTING",
         }
 
-        engine = Engine.from_config(
+        engine = OTXEngine.from_config(
             config_path=recipe_path,
             data_root=data_root,
             task=task_type,
@@ -421,6 +421,6 @@ class TestEngine:
         assert fxt_engine._cache.args.get("devices") == 2
 
         data_root = "tests/assets/classification_dataset"
-        engine = Engine(work_dir=tmp_path, data_root=data_root, num_devices=3)
+        engine = OTXEngine(work_dir=tmp_path, data_root=data_root, num_devices=3)
         assert engine.num_devices == 3
         assert engine._cache.args.get("devices") == 3
